add new scheme entity
