+++ app/common/containers/blocks/Nav/index.js
@@ -52,7 +52,7 @@ export default class Nav extends React.Component {
               <Link id="employees-nav" to="/employees">{ t('Employees') }</Link>
             </NavItem>
           </ShowWithScope>
-          <ShowWithScope scope="employee:read">
+          <ShowWithScope scope="employee_request:read">
             <NavItem to="pending-employees" activeClassName={styles.active}>
               <Link id="pending-employees-nav" to="/pending-employees">{ t('Pending employees') }</Link>
             </NavItem>
