+++ app/common/containers/pages/PendingEmployeesListPage/redux.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import { handleAction, createAction } from 'redux-actions';
-import * as fromEmployees from 'redux/employees';
+import * as fromEmployees from 'redux/employees-requests';
 
 export const showPendingEmployees = createAction('pendingEmployeesListPage/SHOW_DECLARATIONS');
 export const pagingEmployees = createAction('pendingEmployeesListPage/ADD_PAGING');
@@ -17,10 +17,10 @@ export const fetchEmployeesRequest = options => dispatch =>
     return action;
   });
 
-const employees = handleAction(showPendingEmployees, (state, action) => action.payload, []);
+const employeesRequests = handleAction(showPendingEmployees, (state, action) => action.payload, []);
 const paging = handleAction(pagingEmployees, (state, action) => action.payload, {});
 
 export default combineReducers({
-  employees,
+  employeesRequests,
   paging,
 });
