+++ app/common/redux/employees.js
@@ -34,44 +34,12 @@ export const fetchEmployee = id => invoke({
   }, 'employees/FETCH_DETAILS_FAILURE'],
 });
 
-export const fetchEmployeesRequest = options => invoke({
-  endpoint: createUrl(`${API_URL}/api/employee_requests`, options),
-  method: 'GET',
-  headers: {
-    'content-type': 'application/json',
-  },
-  types: ['employees/FETCH_REQUEST_LIST_REQUEST', {
-    type: 'employees/FETCH_REQUEST_LIST_SUCCESS',
-    payload: (action, state, res) => res.clone().json().then(
-      json => normalize(json.data, [employee])
-    ),
-    meta: (action, state, res) => res.clone().json().then(json => json.paging || { cursors: {} }),
-  }, 'employees/FETCH_REQUEST_LIST_FAILURE'],
-});
-
-export const fetchEmployeeRequest = id => invoke({
-  endpoint: createUrl(`${API_URL}/api/employee_requests/${id}`),
-  method: 'GET',
-  headers: {
-    'content-type': 'application/json',
-  },
-  types: ['employees/FETCH_REQUEST_DETAILS_REQUEST', {
-    type: 'employees/FETCH_REQUEST_DETAILS_SUCCESS',
-    payload: (action, state, res) => res.json().then(
-      json => normalize(json.data, employee)
-    ),
-  }, 'employees/FETCH_REQUEST_DETAILS_FAILURE'],
-});
-
 export default handleAction(
   combineActions(
     'employees/FETCH_LIST_SUCCESS',
     'employees/FETCH_DETAILS_SUCCESS',
     'employees/CREATE_SUCCESS',
     'employees/UPDATE_SUCCESS',
-    'employees/FETCH_REQUEST_LIST_SUCCESS',
-    'employees/FETCH_REQUEST_DETAILS_SUCCESS',
-
   ),
   (state, action) => ({
     ...state,
