+++ app/common/schemas/index.js
@@ -6,6 +6,7 @@ export const dictionary = new schema.Entity('dictionaries', {}, { idAttribute: '
 export const clinic = new schema.Entity('clinics');
 export const declaration = new schema.Entity('declarations');
 export const employee = new schema.Entity('employees');
+export const employeesRequest = new schema.Entity('employeesRequests');
 
 export const globalStat = new schema.Entity('globalStat');
 export const detailStat = new schema.Entity('detailStat');
